Refactor ELB sycning to support a second implementation
