+++ BaragonCore/src/main/java/com/hubspot/baragon/models/TrafficSourceType.java
@@ -1,0 +1,5 @@
+package com.hubspot.baragon.models;
+
+public enum TrafficSourceType {
+  CLASSIC, APPLICATION, NETWORK
+}
