+++ BaragonService/src/main/java/com/hubspot/baragon/service/elb/ReigsterInstanceResult.java
@@ -1,0 +1,5 @@
+package com.hubspot.baragon.service.elb;
+
+public enum ReigsterInstanceResult {
+  ELB_NO_VPC_FOUND, ELB_AND_VPC_FOUND, NOT_FOUND
+}
